Move connectorType() to Plugins
