+++ connect/runtime/src/main/java/org/apache/kafka/connect/runtime/isolation/Plugins.java
@@ -24,6 +24,8 @@ import org.apache.kafka.connect.connector.Connector;
 import org.apache.kafka.connect.connector.Task;
 import org.apache.kafka.connect.errors.ConnectException;
 import org.apache.kafka.connect.runtime.WorkerConfig;
+import org.apache.kafka.connect.runtime.rest.entities.ConnectorType;
+import org.apache.kafka.connect.sink.SinkConnector;
 import org.apache.kafka.connect.storage.Converter;
 import org.apache.kafka.connect.transforms.Transformation;
 import org.slf4j.Logger;
@@ -106,6 +108,27 @@ public class Plugins {
 
     public ClassLoader currentThreadLoader() {
         return Thread.currentThread().getContextClassLoader();
+    }
+
+    /*
+     * Retrieves the connector type by connector name
+     *
+     * @return the ConnectorType for the given connector name.
+     * @throws ConnectException if the ConnectorType cannot be determined
+     */
+    public ConnectorType connectorType(String clsName) {
+        ClassLoader savedLoader = currentThreadLoader();
+        try {
+            savedLoader = compareAndSwapLoaders(clsName);
+            try {
+                return SinkConnector.class.isAssignableFrom(Class.forName(clsName)) ?
+                    ConnectorType.SINK : ConnectorType.SOURCE;
+            } catch (Exception e) {
+                throw new ConnectException("Cannot determine the type for " + clsName);
+            }
+        } finally {
+            Plugins.compareAndSwapLoaders(savedLoader);
+        }
     }
 
     public ClassLoader compareAndSwapWithDelegatingLoader() {
