+++ connect/runtime/src/main/java/org/apache/kafka/connect/runtime/standalone/StandaloneHerder.java
@@ -111,7 +111,7 @@ public class StandaloneHerder extends AbstractHerder {
             return null;
         Map<String, String> config = configState.connectorConfig(connector);
         return new ConnectorInfo(connector, config, configState.tasks(connector),
-            worker.connectorType(configState.connectorConfig(connector).get(ConnectorConfig.CONNECTOR_CLASS_CONFIG)));
+            plugins().connectorType(configState.connectorConfig(connector).get(ConnectorConfig.CONNECTOR_CLASS_CONFIG)));
     }
 
     @Override
